feat: enhance validation, error handling and email uniqueness check

+++ src/main/java/ru/basted/corporatedirectory/service/impl/EmployeeServiceImpl.java
@@ -1,5 +1,6 @@
 package ru.basted.corporatedirectory.service.impl;
 
+import ru.basted.corporatedirectory.exception.EmailAlreadyExistsException;
 import ru.basted.corporatedirectory.exception.ResourceNotFoundException;
 import ru.basted.corporatedirectory.model.Employee;
 import ru.basted.corporatedirectory.repository.EmployeeRepository;
@@ -24,18 +25,32 @@ public class EmployeeServiceImpl implements EmployeeService {
     @Override
     public Employee getEmployeeById(Long id) {
         return repository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not found with id: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Сотрудник c id " + id + " не найден"));
     }
 
     @Override
     public Employee createEmployee(Employee employee) {
+        if (repository.existsByEmail(employee.getEmail())) {
+            throw new EmailAlreadyExistsException(
+                    "Пользователь с email " + employee.getEmail() + " уже существует"
+            );
+        }
+
         return repository.save(employee);
     }
 
     @Override
     public Employee changeEmployee(Long id, Employee employee) {
         Employee existingEmployee = repository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not found with id: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Сотрудник c id " + id + " не найден"));
+
+        String newEmail = employee.getEmail();
+
+        if (!existingEmployee.getEmail().equals(newEmail) && repository.existsByEmail(newEmail)) {
+            throw new EmailAlreadyExistsException(
+                    "Пользователь с email " + employee.getEmail() + " уже существует"
+            );
+        }
 
         existingEmployee.setFullName(employee.getFullName());
         existingEmployee.setEmail(employee.getEmail());
@@ -47,6 +62,10 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public void removeEmployee(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Сотрудник c id " + id + " не найден");
+        }
+
         repository.deleteById(id);
     }
 }
